requirements.txt
pyspark==3.5.1
PyYAML==6.0
python-dotenv==1.0.1
kaggle==1.6.17
